requirements.txt
flask
Flask-Session
gunicorn
requests
openai>=1.0.0
pygame
gtts
pytz
openai-whisper
python-docx
ffmpeg-python
moviepy
pillow
pypdf2
werkzeug
imageio
imageio-ffmpeg
mercadopago==2.3.0
requests==2.31.0
python-dateutil
